Refactor DTO imports and update LoginService to use nickname for authentication
