+++ Backend/src/main/java/com/pg/mbti/controllers/UsersController.java
@@ -2,6 +2,8 @@ package com.pg.mbti.controllers;
 
 import com.pg.mbti.entity.user.User;
 import com.pg.mbti.services.UsersService;
+import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,14 +11,11 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
+@AllArgsConstructor
 @RequestMapping("/api")
 public class UsersController {
 
     UsersService usersService;
-
-    UsersController(UsersService usersService) {
-        this.usersService = usersService;
-    }
 
     @GetMapping("/users")
     public List<User> getUsers() {
